ToCatQueue: suspend recursive calls of linkAll
As noted by Okasaki (page 96), one should suspend recursive calls to `linkAll` in order to 'archieve good amortized bounds even in the face of persistence'. Previous implementation called `linkAll` eagerly. Since it was originally assumed in Okasaki's work that `linkAll` is called on a non-empty queue, new implementation explicitly sends head and tail of a queue, thereby statically ensuring precondition. Head is forced right after a call, as in Okasaki's definition. One can use `-ddump-simpl` or `-ddump-stg` GHC flags to verify that recursive call to `linkAll` is enveloped in a thunk. That's because `(|>)` is (assumed to be) non-strict in its second argument. Moreover, thanks to ScopedTypeVariables `TASequence` constraint has disappeared from `linkAll` - dictionary is taken from enclosing `tviewl` context.